# specs/001-notebook-management-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Notebook Management API
  description: REST API for managing notebooks, sections, pages, and tags in a hierarchical note-taking system
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://notebooks.example.com
    description: Production server
...
paths:
...
  /api/notebooks/{notebookId}/sections:
    get:
      summary: List sections in a notebook
      tags: [sections]
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Section'
    
    post:
      summary: Create a new section
      tags: [sections]
      parameters:
        - $ref: '#/components/parameters/NotebookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSectionRequest'
      responses:
        '201':
          description: Section created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Section'
                  message:
                    type: string
  /api/sections/{sectionId}:
    get:
      summary: Get section by ID
      tags: [sections]
      parameters:
        - $ref: '#/components/parameters/SectionId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Section'
    
    put:
      summary: Update section
      tags: [sections]
      parameters:
        - $ref: '#/components/parameters/SectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSectionRequest'
      responses:
        '200':
          description: Section updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Section'
    
    delete:
      summary: Delete section
      tags: [sections]
      parameters:
        - $ref: '#/components/parameters/SectionId'
      responses:
        '200':
          description: Section deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/sections/{sectionId}/reorder:
    post:
      summary: Reorder sections
      tags: [sections]
      description: Update display order of sections within a notebook
      parameters:
        - $ref: '#/components/parameters/SectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_order:
                  type: integer
                  description: New display order position
      responses:
        '200':
          description: Section reordered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Page Endpoints
  /api/sections/{sectionId}/pages:
    get:
      summary: List pages in a section
      tags: [pages]
      parameters:
        - $ref: '#/components/parameters/SectionId'
        - name: include_subpages
          in: query
          schema:
            type: boolean
            default: false
          description: Include nested subpages in response
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Page'
    
    post:
      summary: Create a new page
      tags: [pages]
      parameters:
        - $ref: '#/components/parameters/SectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
      responses:
        '201':
          description: Page created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Page'
                  message:
                    type: string
  /api/pages/{pageId}:
    get:
      summary: Get page by ID
      tags: [pages]
      parameters:
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PageDetail'
    
    put:
      summary: Update page
      tags: [pages]
      description: Update page title, content, or tags
      parameters:
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePageRequest'
      responses:
        '200':
          description: Page updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Page'
    
    delete:
      summary: Delete page
      tags: [pages]
      parameters:
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: Page deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/tags/{tagId}/pages:
    get:
      summary: Get all pages with a specific tag
      tags: [tags]
      parameters:
        - $ref: '#/components/parameters/TagId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Page'

  # Search Endpoints
...
components:
  parameters:
    NotebookId:
      name: notebookId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the notebook
    SectionId:
      name: sectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the section
    PageId:
      name: pageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the page
    TagId:
      name: tagId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the tag
  schemas:
...
    Section:
      type: object
      properties:
        id:
          type: string
          format: uuid
        notebook_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        display_order:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        page_count:
          type: integer
    CreateSectionRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    UpdateSectionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        display_order:
          type: integer
          minimum: 0
    Page:
      type: object
      properties:
        id:
          type: string
          format: uuid
        section_id:
          type: string
          format: uuid
        parent_page_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          minLength: 1
          maxLength: 255
        preview:
          type: string
          description: First 200 characters of content
        display_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    PageDetail:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: string
              description: Full markdown content
            content_html:
              type: string
              description: Rendered HTML content
            subpages:
              type: array
              items:
                $ref: '#/components/schemas/Page'
    CreatePageRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        content:
          type: string
          default: ''
        parent_page_id:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
          description: Array of tag names
    UpdatePageRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        content:
          type: string
        tags:
          type: array
          items:
            type: string
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 50
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        created_at:
          type: string
          format: date-time
        page_count:
          type: integer
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...